yeni inka

+++ database/migrations/2025_06_13_105515_add_reference_on_user_pkl_siswa.php
@@ -14,8 +14,11 @@ class AddReferenceOnUserPklSiswa extends Migration
     public function up()
     {
         Schema::table('users', function (Blueprint $table) {
-            $table->unsignedBigInteger('pkl_id')->nullable();
-            $table->foreign('pkl_id')->references('id')->on('pkls')->onDelete('cascade');
+            // Tambah kolom hanya jika belum ada
+            if (!Schema::hasColumn('users', 'pkl_id')) {
+                $table->unsignedBigInteger('pkl_id')->nullable();
+                $table->foreign('pkl_id')->references('id')->on('pkls')->onDelete('cascade');
+            }
         });
     }
 
@@ -26,6 +29,12 @@ class AddReferenceOnUserPklSiswa extends Migration
      */
     public function down()
     {
-        //
+        Schema::table('users', function (Blueprint $table) {
+            // Hapus kolom hanya jika ada
+            if (Schema::hasColumn('users', 'pkl_id')) {
+                $table->dropForeign(['pkl_id']);
+                $table->dropColumn('pkl_id');
+            }
+        });
     }
 }
